% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed-insights-package.R
\name{speedfinder2}
\alias{speedfinder2}
\title{Function within speedlist}
\usage{
speedfinder2(url, strategy, key)
}
\description{
Speedfinder2 is a function utilized by speedlist to allow a list of URLs to be submitted to the Google Page Speed
Insights API.
}
