% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed-insights-package.R
\name{speedlist}
\alias{speedlist}
\title{Speed results for a list of URLs}
\usage{
speedlist(pagelist, strategy, key)
}
\arguments{
\item{strategy}{Whether the list of URLs should be evaluated in a mobile or desktop context. Accordingly, acceptable values
are either the string "mobile" or the string "desktop".}

\item{key}{A unique key obtainable from Google by registering for free as a Google developer.}

\item{url}{A list of URLs on which to run the Google Page Speed Insights test.}
}
\description{
The speedlist function submits a list of URLs to the Google Page Speed Insights API, and returns a dataframe in which
each row contains the results from the API for a unique URL in the list.
}
\examples{
speedlist(listofURLs,"mobile",key)
}
